Remove drop check from covariant type parameters.
This change replaces `PhantomData<T>` fields with `PhantomData<fn() ->
T>` and `PhantomData<*const T>` fields. All are covariant with respect
to `T`, but the latter do not require a drop check. The affected types
never own instances of these type parameters, so a drop check is not
necessary.

Moreover, this change removes unneeded bounds on `Constraint` and
unneeded implementations of `Clone` and `Copy` for constraint types.
`Proxy` now implements `Clone` and `Copy` with no bounds on its
constraint type `P`.