Introduce configurable error behavior with divergence.
This change introduces components that configure the output types and
behavior of `Proxy` types when performing fallible operations.
Constraints now have an associated divergence, which decides what
happens when a constraint `Result` is `Ok` or `Err`. A divergence
composes an output kind, which determines the output type when
diverging.

These types can be composed to achieve a desired output type and
behavior. For example:

```
pub type Real = Proxy<f64, FiniteConstraint<OrPanic<AsSelf>>>;
```

This `Real` type is an `f64` that returns `Real` (`Self`) in fallible
operations and panics if any of these operations fail (do not produce a
real number).

To support this change, a `real` module has been introduced with traits
describing types that support real functions. An `expression` module and
`Expression` type have also been introduced, which support fallible
outputs that can be used in expressions to defer error handling.
Finally, this change upgrades to Rust 2021 Edition and bumps the MSRV to
`1.65.0`.