Use more explicit API for the sign of real numbers.
This change replaces the typical `is_positive`, `is_negative`, etc. APIs
with a `UnaryReal::sign` function and corresponding `Sign` type. `Sign`
represents three states: positive, negative, and zero. The zero sign is
the neutral sign for zero in the real numbers but is also used for
floating-point implementations for `NaN`, which also has a neutral sign
(neither positive nor negative).