Rename and remove some traits in the `real` module.
This change renames traits in the `real` module such that they emphasize
constants and functions over reals and avoid the suggestion that a bound
like `T: RealFunction` (`T: Real` before this change) means that `T`
**is** a "real number type". Rather, this bound (merely) means that `T`
has real inhabitants and supports real functions; such a `T` may **not**
represent a real number.

This change also removes the `ExtendedReal` trait. This trait had the
same problem as above, but additionally involved a non-real trait more
concerned with IEEE 754 floating-point and can be derived trivially if
needed: `Infinity + RealFunction`.