Remove `Float` and rename IEEE 754 traits.
This change renames IEEE 754 traits such that they describe encoding.
The `Encoding` trait is now `BaseEncoding`, and `Nan` is now
`NanEncoding`, for example. This supports upcoming types and names that
describe these encodings and makes it clearer that these traits are
related.

Importantly, this change removes the `Float` trait. This trait was an
"alias" (blanket implemented for types that implement its supertraits)
and is probably less useful than it appears. Most code should rely on a
more defined type like `Proxy<T, IsFloat>` or just `R64` or, if bounds
are needed, something like `T: FloatFunction<f64>` or `T:
RealEndofunction`.

If it turns out that removing `Float` is a mistake, it can be spelled
using other traits in this crate in the meantime. If reintroduced,
`FloatEncoding` is likely a good name, but this may imply features that
this trait does not specify, such as equivalence relations and ordering.
Note that both `f64` and `Total<f64>` implemented this trait! The former
is compliant with IEEE 754 and the latter is not.