Remove the notion of "bits" from canonicalization.
This change renames `ToCanonicalBits` to `ToCanonical`. Associated items
like types and functions are also renamed accordingly, such as
`CanonicalEq::eq_canonical_bits` to `CanonicalEq::eq_canonical`.

This naming is a bit more consistent for ordering and the `CanonicalOrd`
trait, which while related, is not implemented in terms of `ToCanonical`
and does not compare its outputs nor bits. Put another way, the name
`CanonicalOrd::cmp_canonical_bits` is misleading.