Improve `Debug` formatting for proxy types.
This change introduces the `Proxy::fmt` and `Proxy::debug` functions,
which provide thorough debugging descriptions of proxy types. These
descriptions include the constraint and divergence, unlike `Debug`
implementations for `Proxy`, which are intentionally more terse.

This change also broadens the `Debug` implementations for `Proxy`,
instead basing them on a `Proxy` type with a particular `Constraint`
type rather than on the crate-level type definitions, which are more
specific.