Refactor `IntrinsicOrd`.
This change reimplements the `IntrinsicOrd` trait and reworks its basic
mechanisms. `IntrinsicOrd` now more closely resembles `PartialOrd` and
`Ord`, in that types are queried for an `Ordering` (rather than directly
providing functions like `max` that order two or more values). These
queries return a `Result`, where the error variant contains a notion of
`Undefined` for the type.

For partially ordered types with a notion of undefined, the `Undefined`
type is inhabited and is typically `Self`. For totally ordered types
with no intrinsic representation of undefined, this type is uninhabited.
Functions for constructing an `IntrinsicOrd` type from `Undefined` then
perform any necessary conversion or become unreachable, respectively.