Replace terms "intrinsic" and "undefined" with "empty".
This change renames `IntrinsicOrd` and related APIs. The terms
"intrinsic" and "undefined" are dropped in favor of "empty". In
particular, this notion of undefined can be confusing, because it is
unrelated to constraints. Though the notion of empty inhabitants is a
bit more obtuse, it is less likely to be confused with constraints,
codomains, etc.